refactor: moving background functions and logic
