+++ src/scripts/background/alert/setUninstalledRedirect.js
@@ -1,0 +1,6 @@
+export function setUninstalledRedirect() {
+  chrome.runtime.setUninstallURL(
+    'https://save.hugocollin.com/uninstalled',
+    () => console.log('Uninstall survey URL set')
+  );
+}
