+++ src/scripts/background/background.js
@@ -1,8 +1,9 @@
-import {listenIconClick} from "./chrome/icon/iconAction";
-import {buildContextMenu} from "./chrome/buildContextMenu";
-import {listenTabsToUpdateIcon} from "./chrome/icon/defineIcon";
-import {initClickCount} from "./chrome/icon/clickCount";
-import {initModalOnInstall, setUninstalledRedirect} from "./chrome/alert";
+import {listenIconClick} from "./icon/iconAction";
+import {buildContextMenu} from "./contextMenu/buildContextMenu";
+import {listenTabsToUpdateIcon} from "./icon/defineIcon";
+import {initClickCount} from "./icon/clickCount";
+import {initModalOnInstall} from "./alert/manageModals";
+import {setUninstalledRedirect} from "./alert/setUninstalledRedirect";
 // import {launchScrappingActionOnPage} from "../../content/launch/launchScraperOnPage";
 // import {defineProcessingState} from "../scraper/defineProcessingState";
 
@@ -15,6 +16,10 @@ function background() {
   initClickCount();
   initModalOnInstall();
   setUninstalledRedirect();
+
+  chrome.runtime.onInstalled.addListener(() => {
+    chrome.storage.sync.set({filenameFormat: '%t'});
+  });
 
   // exportAllThreadsListener();
   // scrapOnLoadListener();
